# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/chat": {
      "get": {
        "description": "Retrieves a paginated list of all chats where the user is either admin or participant",
        "operationId": "ChatController_getChats",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "description": "Filter chats by project ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of chats to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of chats to skip (default: 0)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status (default: true)",
            "schema": {
              "example": "true",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all chats for the authenticated user",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/project/{projectId}": {
      "get": {
        "description": "Retrieves a paginated list of all chats in a specific project",
        "operationId": "ChatController_getProjectChats",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "adminId",
            "required": false,
            "in": "query",
            "description": "Filter chats by admin ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": false,
            "in": "query",
            "description": "Filter chats by participant ID",
            "schema": {
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of chats to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of chats to skip (default: 0)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved project chats",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all chats for a specific project",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/{chatId}": {
      "get": {
        "description": "Retrieves information about a specific chat",
        "operationId": "ChatController_getChatById",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved chat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          },
          "403": {
            "description": "User does not have access to this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific chat by ID",
        "tags": [
          "chat"
        ]
      },
      "delete": {
        "description": "Deactivates a chat (soft delete)",
        "operationId": "ChatController_deactivateChat",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deactivated chat"
          },
          "401": {
            "description": "User is not authenticated"
          },
          "403": {
            "description": "User does not have access to this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a chat",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/project/{projectId}/participant/{participantId}": {
      "post": {
        "description": "Creates a new chat or returns existing one between admin and participant",
        "operationId": "ChatController_createOrGetChat",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "ID of the project",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "participantId",
            "required": true,
            "in": "path",
            "description": "ID of the participant",
            "schema": {
              "example": 2,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved or created chat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or get chat with participant",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/{chatId}/messages": {
      "get": {
        "description": "Retrieves a paginated list of messages for a specific chat",
        "operationId": "ChatController_getMessages",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Maximum number of messages to return (default: 20)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of messages to skip (default: 0)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved messages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedMessagesResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          },
          "403": {
            "description": "User does not have access to this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get messages for a chat",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "description": "Creates a new message in a specific chat",
        "operationId": "ChatController_createMessage",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageViaApiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid message content or spam protection triggered"
          },
          "401": {
            "description": "User is not authenticated"
          },
          "403": {
            "description": "User does not have access to this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a message to a chat",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/{chatId}/read": {
      "put": {
        "description": "Marks all messages in a chat as read for the authenticated user",
        "operationId": "ChatController_markChatAsRead",
        "parameters": [
          {
            "name": "chatId",
            "required": true,
            "in": "path",
            "description": "ID of the chat",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully marked chat as read"
          },
          "401": {
            "description": "User is not authenticated"
          },
          "403": {
            "description": "User does not have access to this chat"
          },
          "404": {
            "description": "Chat not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark chat as read",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/status/{userId}": {
      "get": {
        "description": "Retrieves the online status of a user",
        "operationId": "ChatController_getUserOnlineStatus",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "path",
            "description": "ID of the user to get status for (defaults to authenticated user)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved online status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isOnline": {
                      "type": "boolean",
                      "example": true
                    },
                    "lastSeen": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-01-15T10:00:00.000Z",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user online status",
        "tags": [
          "chat"
        ]
      }
    },
    "/api/chat/status": {
      "get": {
        "description": "Retrieves the online status of the authenticated user",
        "operationId": "ChatController_getCurrentUserOnlineStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved online status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isOnline": {
                      "type": "boolean",
                      "example": true
                    },
                    "lastSeen": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2025-01-15T10:00:00.000Z",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User is not authenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user online status",
        "tags": [
          "chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat CRM API",
    "description": "API documentation for Chat CRM application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "ProjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Project name",
            "example": "My Project"
          },
          "uniqueId": {
            "type": "string",
            "description": "Project unique ID",
            "example": "proj123"
          }
        },
        "required": [
          "id",
          "name",
          "uniqueId"
        ]
      },
...
      "ChatResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Chat ID",
            "example": 1
          },
          "projectId": {
            "type": "number",
            "description": "Project ID",
            "example": 1
          },
          "project": {
            "description": "Project information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ProjectDto"
              }
            ]
          },
          "admin": {
            "description": "Admin user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatParticipantDto"
              }
            ]
          },
          "participant": {
            "description": "Participant user information",
            "allOf": [
              {
                "$ref": "#/components/schemas/ChatParticipantDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Chat creation timestamp",
            "example": "2025-01-15T10:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Chat last update timestamp",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the chat is active",
            "example": true
          },
          "lastMessage": {
            "description": "Last message in the chat",
            "allOf": [
              {
                "$ref": "#/components/schemas/LastMessageDto"
              }
            ]
          },
          "unreadCount": {
            "type": "number",
            "description": "Number of unread messages for the requesting user",
            "example": 3
          },
          "lastMessageAt": {
            "format": "date-time",
            "type": "string",
            "description": "Timestamp of the last message",
            "example": "2025-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "projectId",
          "project",
          "admin",
          "participant",
          "createdAt",
          "updatedAt",
          "isActive",
          "unreadCount"
        ]
      },
...
      "PaginatedMessagesResponseDto": {
        "type": "object",
        "properties": {
          "messages": {
            "description": "Array of messages",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageResponseDto"
            }
          },
          "totalCount": {
            "type": "number",
            "description": "Total number of messages in the chat",
            "example": 50
          },
          "limit": {
            "type": "number",
            "description": "Number of messages per page",
            "example": 20
          },
          "offset": {
            "type": "number",
            "description": "Number of messages skipped",
            "example": 0
          }
        },
        "required": [
          "messages",
          "totalCount",
          "limit",
          "offset"
        ]
      },
      "CreateMessageViaApiDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "description": "Content of the message",
            "example": "Hello! How can I help you today?",
            "maxLength": 1000
          },
          "type": {
            "type": "string",
            "description": "Type of the message. Use IMAGE for image files, FILE for other files, TEXT for text messages",
            "enum": [
              "TEXT",
              "EMOJI",
              "FILE",
              "IMAGE",
              "SYSTEM"
            ],
            "example": "TEXT",
            "default": "TEXT"
          },
          "metadata": {
            "type": "object",
            "description": "Additional metadata for the message (e.g., file info, emoji data)",
            "example": {
              "fileUrl": "https://storage.googleapis.com/bucket/chat/1/uuid.jpg",
              "fileName": "image.jpg",
              "fileSize": 1024000,
              "mimeType": "image/jpeg",
              "originalName": "photo.jpg"
            }
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}